Refactor using bulletproof react
https://github.com/alan2207/bulletproof-react/blob/master/docs/project-structure.md
